split RandomizeCharacters and created GetImagesHP component
